+++ kibot/globals.py
@@ -100,6 +100,8 @@ class Globals(FiltersOptions):
             self.copper_thickness = 35
             """ [number|string] Copper thickness in micrometers (1 Oz is 35 micrometers).
                 KiCad 6: you should set this in the Board Setup -> Physical Stackup """
+            self.cross_no_body = False
+            """ Cross components even when they don't have a body. Only for KiCad 6 """
             self.date_format = '%Y-%m-%d'
             """ Format used for the day we started the script.
                 Is also used for the PCB/SCH date formatting when `time_reformat` is enabled (default behavior).