+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Try to download missing tools and Python modules.
   The user also gets more information when something is missing.
   It can be disabled from the command line.
-
+- A global option to cross components without a body.
 
 ## [1.2.0] - 2022-06-15
 ### Added